all authentication and pundit roles are set up. need to clean slate and delete all duplicate models and prepare for presentation before class. everything works perfect, all data is written in file and will be transferred to notes on my phone or a google doc or sheet. will ru a second test beforehand tomorrow before class to ensure there are no errors mid presentation.

+++ app/controllers/cars_controller.rb
@@ -1,31 +1,44 @@
 class CarsController < ApplicationController
   before_action :authorize_request
   def index # use get
-    cars = Car.all 
+    cars = policy_scope(Car)
     render json: cars
   end
 
   def show # use get
     car = Car.find(params[:id]) #dynamic
+    authorize car  # calls CarPolicy#show?
     render json: car
   end
 
   def create # use post
-    car = Car.create!(model:params[:model], year:params[:year],make_id:params[:make_id]) #dynamic     
-    render json: car
+    car = Car.new(model:params[:model], year:params[:year],make_id:params[:make_id],  user_id: current_user.id)    
+    
+    authorize car  # calls CarPolicy#create?
+
+    if car.save
+      render json: car, status: :created
+    else
+      render json: { errors: car.errors.full_messages }, status: :unprocessable_entity
+    end
   end
 
   def update # use put or patch
-    car = Car.find(params[:id]) # dynamic
-    car.update(model:
-    params[:model], year:params[:year],make_id:params[:make_id]
-    )
-    render json: car
+    car = Car.find(params[:id]) #dynamic
+    authorize car  # calls carPolicy#update?
+
+    if car.update(model: params[:model], year: params[:year],make_id: params[:make_id])
+      render json: car
+    else
+      render json: { errors: car.errors.full_messages }, status: :unprocessable_entity
+    end 
   end
   
   def destroy # use delete
     car = Car.find(params[:id]) # dynamic
+    authorize car  # calls CarPolicy#destroy?
+
     car.destroy
-    render :no_content
+    render json: { message: 'Car has been deleted.' }, status: :ok
   end
 end
